requirements.txt
asgiref==3.5.2
astroid==2.12.13
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.4
PyJWT==2.6.0
pylint==2.15.6
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
